fix: added [slug] server file for ssr data fetching
