+++ .env
@@ -1,0 +1,1 @@
+VITE_ANALOG_PUBLIC_BASE_URL="http://localhost:5173"
